refactor: extract about feature
