+++ package.json
@@ -12,6 +12,7 @@
     "test:aviation-data": "node src/services/aviationData.test.js",
     "test:vercel-routing": "node src/config/vercelRouting.test.js",
     "test:airport-wiki": "node src/services/airportWiki.test.js",
+    "test:about": "node src/features/about/aboutModel.test.js",
     "test:app-shell": "node src/features/app-shell/themePreference.test.js",
     "test:airport-search": "node src/features/airport-search/airportSearchModel.test.js",
     "test:airport-explorer": "node src/features/airport-explorer/airportExplorerModel.test.js",
